유진호 06/19 12:10
종목 상세보기 페이지 뉴스

+++ F5/src/main/java/com/smhrd/stock/entity/NewsForCompany.java
@@ -22,7 +22,7 @@ public class NewsForCompany {
 	private String stockCode;
 
 	@Column(name="news_idx") 
-	private int newsIdx;
+	private Long newsIdx;
 	
 	@Column(name="relevance_score", precision = 5, scale = 2)
 	private BigDecimal relevanceScore;
